Use dbus-run-session to run D-Bus tests instead of dbus-launch
dbus-run-session spawns a session that gets terminated automatically
after the execution of the command. It is a simpler way compared to
dbus-launch, which relies on the X11 tolaunching (which would explain
why it is not used on macOS, as the comment says).

Hence, use dbus-run-session instead of dbus-launch to run each D-Bus
test. Keep the existing macOS behaviour as-is for now, adding a comment
to test using dbus-run-session in the future.